+++ packages/nimble-components/src/dialog/tests/dialog-matrix.stories.ts
@@ -1,10 +1,19 @@
 import type { StoryFn, Meta } from '@storybook/html';
-import { html } from '@microsoft/fast-element';
+import { html, ViewTemplate } from '@microsoft/fast-element';
 import { createFixedThemeStory } from '../../utilities/tests/storybook';
 import { sharedMatrixParameters } from '../../utilities/tests/matrix';
 import { backgroundStates } from '../../utilities/tests/states';
 import { dialogTag } from '..';
 import { buttonTag } from '../../button';
+import {
+    bodyFont,
+    dialogLargeHeight,
+    dialogLargeMaxHeight,
+    dialogLargeWidth,
+    dialogSmallHeight,
+    dialogSmallMaxHeight,
+    dialogSmallWidth
+} from '../../theme-provider/design-tokens';
 
 const metadata: Meta = {
     title: 'Tests/Dialog',
@@ -15,9 +24,39 @@ const metadata: Meta = {
 
 export default metadata;
 
+const sizeStates = [
+    `width: var(${dialogSmallWidth.cssCustomProperty}); height: var(${dialogSmallHeight.cssCustomProperty}); max-height: var(${dialogSmallMaxHeight.cssCustomProperty});`,
+    `width: var(${dialogLargeWidth.cssCustomProperty}); height: var(${dialogLargeHeight.cssCustomProperty}); max-height: var(${dialogLargeMaxHeight.cssCustomProperty});`
+] as const;
+type SizeState = (typeof sizeStates)[number];
+
 const component = html`
     <${dialogTag}>
         <span slot="title">This is my dialog's title. It is pretty long.</span>
+        <span slot="subtitle">The dialog has a subtitle here.</span>
+        <div>Here is the first piece of content in the dialog</div>
+        <div>
+            Here is another piece of content in the dialog. It is a bit longer.
+        </div>
+        <${buttonTag} slot="footer">Cancel</${buttonTag}>
+        <${buttonTag} slot="footer">OK</${buttonTag}>
+    </${dialogTag}>
+`;
+
+const dialogSizingTestCase = (size: SizeState): ViewTemplate => html`
+    <p class="spacer">${() => size};</p>
+    <style>
+        ${dialogTag}::part(control) {
+            ${() => size};
+        }
+
+        .spacer {
+            font: var(${bodyFont.cssCustomProperty});
+            padding-bottom: 1000px;
+        }
+    </style>
+    <${dialogTag}>
+    <span slot="title">This is my dialog's title. It is pretty long.</span>
         <span slot="subtitle">The dialog has a subtitle here.</span>
         <div>Here is the first piece of content in the dialog</div>
         <div>
@@ -60,3 +99,17 @@ export const dialogDarkThemeBlackBackground: StoryFn = createFixedThemeStory(
 );
 
 dialogDarkThemeBlackBackground.play = playFunction;
+
+export const dialogSmallSize: StoryFn = createFixedThemeStory(
+    dialogSizingTestCase(sizeStates[0]),
+    lightThemeWhiteBackground
+);
+
+dialogSmallSize.play = playFunction;
+
+export const dialogLargeSize: StoryFn = createFixedThemeStory(
+    dialogSizingTestCase(sizeStates[1]),
+    lightThemeWhiteBackground
+);
+
+dialogLargeSize.play = playFunction;